requirements.txt
absl-py==2.1.0
accelerate==1.1.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alias-free-torch==0.0.6
aliyun-python-sdk-core==2.15.2
aliyun-python-sdk-kms==2.16.5
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrdict==2.0.1
attrs==24.2.0
audioread==3.0.1
bert-score==0.3.13
binpacking==1.5.2
bitarray==3.0.0
bitstring==4.3.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cocoNLP==0.0.13
coloredlogs==15.0.1
contourpy==1.3.0
cpm-kernels==1.0.11
crcmod==1.7
cryptography==43.0.1
cycler==0.12.1
dacite==1.8.1
datasets==3.2.0
decorator==5.1.1
dill==0.3.8
Distance==0.1.3
distro==1.9.0
Django==5.1.1
docstring_parser==0.16
dotenv==0.9.9
editdistance==0.5.2
einops==0.8.0
espnet-tts-frontend==0.0.3
et_xmlfile==2.0.0
exceptiongroup==1.2.2
Faker==36.2.2
fastapi==0.114.1
ffmpy==0.4.0
filelock==3.13.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.2.0
funasr==1.2.6
funcodec==0.2.0
future==1.0.0
g2p==2.2.0
g2p-en==2.1.0
gradio==4.31.3
gradio_client==0.16.3
greenlet==3.1.1
grpcio==1.66.1
h11==0.14.0
h5py==3.13.0
hanlp_downloader==0.0.25
hanziconv==0.3.2
hdbscan==0.8.40
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.30.2
humanfriendly==10.0
hydra-core==1.3.2
hyper-connections==0.1.15
HyperPyYAML==1.2.2
idna==3.8
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflect==7.0.0
jaconv==0.4.0
jamo==0.4.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.3
jiter==0.5.0
jmespath==0.10.0
joblib==1.4.2
jpype1==1.5.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kaldiio==2.18.1
kantts==1.0.1
kiwisolver==1.4.7
kwsbp==0.1.0
lazy_loader==0.4
librosa==0.10.1
liger_kernel==0.5.8
llvmlite==0.43.0
local-attention==1.11.1
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
MinDAEC==0.1.0
mir_eval==0.8.2
modelscope==1.23.2
modelscope_studio==0.4.2
mpmath==1.3.0
ms_swift==3.3.1
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
munkres==1.1.4
narwhals==1.7.0
networkx==3.2.1
ninja==1.11.1.3
nltk==3.9.1
numba==0.60.0
numpy==1.26.3
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.0.44
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu12==12.3.0.142
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.4.99
omegaconf==2.3.0
openai==1.44.1
openpyxl==3.1.5
orjson==3.10.7
oss2==2.19.0
packaging==24.1
pandas==2.2.2
panphon==0.21.2
parso==0.8.4
peft==0.12.0
pexpect==4.9.0
phaseaug==1.0.1
phone==0.4.5
phonenumbers==9.0.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.3.2
pooch==1.8.2
prompt_toolkit==3.0.50
protobuf==5.28.1
psutil==6.0.0
ptflops==0.7.4
ptyprocess==0.7.0
py-sound-connect==0.2.1
pyarrow==17.0.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyhanlp==0.1.89
pynndescent==0.5.13
pyparsing==3.1.4
pypinyin==0.44.0
pysptk==0.1.18
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-wavelets==1.3.0
pytorch-wpe==0.0.1
pytz==2024.2
PyWavelets==1.8.0
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rotary-embedding-torch==0.8.6
rouge==1.0.1
rouge-chinese==1.0.3
rouge-score==0.1.2
rpds-py==0.20.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
ruff==0.6.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
shtab==1.7.1
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
sox==1.5.0
soxr==0.5.0.post1
speechbrain==1.0.2
sqlparse==0.5.1
starlette==0.38.5
sympy==1.12
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
text-unidecode==1.3
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.21.0
tomlkit==0.12.0
torch==2.4.1+cu124
torch-complex==0.4.4
torchaudio==2.4.1+cu124
torchvision==0.19.1+cu124
tqdm==4.66.5
traitlets==5.14.3
transformers==4.51.3
transformers-stream-generator==0.0.5
triton==3.0.0
trl==0.15.2
ttsfrd==0.2.1
typeguard==2.13.3
typer==0.12.5
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tyro==0.8.10
tzdata==2024.1
umap==0.1.1
umap-learn==0.5.7
unicodecsv==0.14.1
Unidecode==1.3.8
urllib3==2.2.2
uvicorn==0.30.6
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.4
xxhash==3.5.0
yarl==1.11.1
zipp==3.20.1
zstandard==0.23.0
